+++ ext/wrapt/__init__.py
@@ -1,10 +1,10 @@
-__version_info__ = ('1', '10', '11')
+__version_info__ = ('1', '14', '1')
 __version__ = '.'.join(__version_info__)
 
 from .wrappers import (ObjectProxy, CallableObjectProxy, FunctionWrapper,
-        BoundFunctionWrapper, WeakFunctionProxy, resolve_path, apply_patch,
-        wrap_object, wrap_object_attribute, function_wrapper,
-        wrap_function_wrapper, patch_function_wrapper,
+        BoundFunctionWrapper, WeakFunctionProxy, PartialCallableObjectProxy,
+        resolve_path, apply_patch, wrap_object, wrap_object_attribute,
+        function_wrapper, wrap_function_wrapper, patch_function_wrapper,
         transient_function_wrapper)
 
 from .decorators import (adapter_factory, AdapterFactory, decorator,
@@ -13,7 +13,15 @@ from .decorators import (adapter_factory, AdapterFactory, decorator,
 from .importer import (register_post_import_hook, when_imported,
         notify_module_loaded, discover_post_import_hooks)
 
-try:
-    from inspect import getcallargs
-except ImportError:
-    from .arguments import getcallargs
+# Import of inspect.getcallargs() included for backward compatibility. An
+# implementation of this was previously bundled and made available here for
+# Python <2.7. Avoid using this in future.
+
+from inspect import getcallargs
+
+# Variant of inspect.formatargspec() included here for forward compatibility.
+# This is being done because Python 3.11 dropped inspect.formatargspec() but
+# code for handling signature changing decorators relied on it. Exposing the
+# bundled implementation here in case any user of wrapt was also needing it.
+
+from .arguments import formatargspec