+++ medusa/name_parser/guessit_parser.py
@@ -109,22 +109,42 @@ def get_expected_titles(show_list):
     """
     expected_titles = []
     for show in show_list:
-        exceptions = {show.name}.union({alias.title for alias in show.aliases})
+        show_title = show.name
+        exceptions = {alias.title for alias in show.aliases}
         for exception in exceptions:
+            # Do not add only numbers to expected titles.
             if exception.isdigit():
-                # do not add numbers to expected titles.
                 continue
 
             match = series_re.match(exception)
             if not match:
                 continue
 
-            if not any(char.isdigit() or char == '-' for char in match.group(1)):
+            # Add when show exception has a year (without brackets),
+            # a number or '-' in its title.
+            if any(char.isdigit() or char == '-' for char in match.group(1)):
+                expected_titles.append(exception)
                 continue
 
-            # At this point only add titles (without the year ex: 9-1-1 (2018),
-            # only use `9-1-1`. And only when it has a number or '-' in its title.
-            expected_titles.append(exception)
+            # Add when show name is the same as exception,
+            # to allow an explicit match.
+            if show_title.casefold() == exception.casefold():
+                expected_titles.append(exception)
+                continue
+
+        # Do not add only numbers to expected titles.
+        if show_title.isdigit():
+            continue
+
+        match = series_re.match(show_title)
+        if not match:
+            continue
+
+        # Add when show exception has a year (without brackets),
+        # a number or '-' in its title.
+        if any(char.isdigit() or char == '-' for char in match.group(1)):
+            expected_titles.append(show_title)
+            continue
 
     return expected_titles